backend/requirements.txt
asgiref==3.7.2
astroid==2.15.8
black==23.9.1
click==8.1.7
dill==0.3.7
Django==4.2.6
djangorestframework==3.14.0
greenlet==2.0.2
input-remapper==2.0.1
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.9
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pynvim==0.4.3
pytz==2023.3.post1
rofi-copyq @ git+https://github.com/cjbassi/rofi-copyq@2ce8628b1e17d91c82d6d40302f1325f3edee207
sqlparse==0.4.4
tomlkit==0.12.1
wrapt==1.15.0
